merge

+++ proj/DynamicObject.h
@@ -16,7 +16,7 @@ public:
 	DynamicObject();
 	~DynamicObject();
 
-	void update(double delta_t);
+	void update(unsigned long delta_t);
 	void setSpeed(Vector3* speed);
 	void setSpeed(double x, double y, double z);
 	Vector3* getSpeed();
